+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Fixed
+- Logs an error if the plugin fails to initialize.
+
 ## [0.5.0] - 2020-02-05
 
 ### Added